authorization delete

+++ src/controllers/postController.js
@@ -1,4 +1,6 @@
 const { getSqlData } = require("../routes/helper");
+const mysql = require("mysql2/promise");
+const { dbConfig } = require("../config");
 
 module.exports.getAll = async (req, res) => {
   // const sql = "SELECT * FROM `posts`";
@@ -42,12 +44,12 @@ module.exports.getAll = async (req, res) => {
 
 module.exports.getSingle = async (req, res, next) => {
   // let connection;
-  const postId = req.params.postId;
+  const { postId } = req.params;
 
   //ar autorizuotas?
 
-  const sql = `SELECT posts.post_id, posts.title, posts.author, users.email AS userEmail, posts.content, posts.date,  COUNT(post_comments.comm_id) AS commentCount,
-  categories.title AS catagoryName
+  const sql = `SELECT posts.post_id, posts.title, posts.author, users.email AS userEmail, posts.content, posts.date, COUNT(post_comments.comm_id) AS commentCount,
+  categories.title AS categoryName
   FROM posts
   JOIN categories
   ON posts.cat_id=categories.cat_id
@@ -97,22 +99,69 @@ module.exports.getSingle = async (req, res, next) => {
   // }
 };
 
-module.exports.delete = async (req, res) => {
+module.exports.delete = async (req, res, next) => {
+  const { postId } = req.params;
+  let conn;
+  try {
+    conn = await mysql.createConnection(dbConfig);
+    // isitikinti kad posto autorius yra tas kuris nori istrinti
+    const sql1 = "SELECT * FROM posts WHERE post_id=?";
+    const [rows1] = await conn.execute(sql1, [postId]);
+
+    if (+rows1[0].user_id !== +req.userId) {
+      next({ status: 401, message: "Not allowed" });
+      return;
+    }
+
+    const sql = "DELETE FROM posts WHERE post_id=? LIMIT 1";
+    const [rows] = await conn.execute(sql, [postId]);
+    console.log("rows ===", rows);
+    // pavyko istrinti jei
+    if (rows.affectedRows === 1) {
+      res.json({ msg: `post with id ${postId} was deleted` });
+      return;
+    }
+    // rows.affectedRows !== 1
+    res.status(400).json({
+      msg: "no rows afected",
+      rows,
+    });
+  } catch (error) {
+    console.warn("DELETE FROM posts error");
+    // res.status(500).json('something wrong');
+    next(error);
+  } finally {
+    // atsijungiam
+    if (conn) conn.end();
+    // conn?.end();
+  }
   // let connection;
-  const postId = req.params.postId;
-  const sql = "DELETE FROM posts WHERE post_id=? LIMIT 1";
-  const [postArr, error] = await getSqlData(sql, [postId]);
-  if (error) {
-    // console.log(error);
-    // res.status(500).json("something wrong");
-    // return;
-    return next(error);
-  }
-  if (postArr.affectedRows === 1) {
-    res.json({ msg: `post with id ${postId} was deleted` });
-    return;
-  }
-  res.status(400).json({ msg: `no rows affected`, postArr });
+  //   const {postId} = req.params
+
+  //   //isitif (req.userEmail)ikinti, kad post autorius yra tas, kuris nori istrinti
+  //   //
+  //   const sql1 = "SELECT * FROM posts WHERE post_id=?";
+  //   const [rows1] = await getSqlData(sql1, [postId]);
+
+  //   if (+rows1[0].user_id !== +req.userId) {
+  //     next({ status: 401, message: "Not allowed" });
+  //     return;
+  //   }
+
+  //   const sql = "DELETE FROM posts WHERE post_id=? LIMIT 1";
+  //   const [postArr, error] = await getSqlData(sql, [postId]);
+
+  //   if (error) {
+  //     // console.log(error);
+  //     // res.status(500).json("something wrong");
+  //     // return;
+  //     return next(error);
+  //   }
+  //   if (postArr.affectedRows === 1) {
+  //     res.json({ msg: `post with id ${postId} was deleted` });
+  //     return;
+  //   }
+  //   res.status(400).json({ msg: `no rows affected`, postArr });
 
   // try {
   //   connection = await mysql.createConnection(dbConfig);
@@ -131,26 +180,58 @@ module.exports.delete = async (req, res) => {
   // }
 };
 
-module.exports.create = async (req, res) => {
+module.exports.create = async (req, res, next) => {
+  console.log("req.body ===", req.body);
   const { title, author, date, content, cat_id: catId } = req.body;
-  const sql = `
-      INSERT INTO posts (title, author, date, content, cat_id)
-      VALUES (?, ?, ?, ?, ?)
+  const { userId } = req;
+
+  // validation
+
+  let conn;
+  try {
+    conn = await mysql.createConnection(dbConfig);
+    const sql = `
+      INSERT INTO posts (title, author, date, content, cat_id, user_id) 
+      VALUES (?,?,?,?,?,?)
       `;
-  const [postArr, error] = await getSqlData(sql, [
-    title,
-    author,
-    date,
-    content,
-    catId,
-  ]);
-  if (error) {
-    // console.log(error);
-    // res.status(500).json("something wrong");
-    // return;
-    return next(error);
+    const [rowOb] = await conn.execute(sql, [
+      title,
+      author,
+      date,
+      content,
+      catId,
+      userId,
+    ]);
+    res.status(201).json(rowOb);
+  } catch (error) {
+    console.warn("INSERT INTO posts err");
+    next(error);
+  } finally {
+    // atsijungiam
+    if (conn) conn.end();
+    // conn?.end();
   }
-  res.json(postArr);
+  //   const { title, author, date, content, cat_id: catId } = req.body;
+  //   const userId = req.userId;
+  //   const sql = `
+  //       INSERT INTO posts (title, author, date, content, cat_id, user_id)
+  //       VALUES (?, ?, ?, ?, ?, ?)
+  //       `;
+  //   const [postArr, error] = await getSqlData(sql, [
+  //     title,
+  //     author,
+  //     date,
+  //     content,
+  //     catId,
+  //     userId,
+  //   ]);
+  //   if (error) {
+  //     // console.log(error);
+  //     // res.status(500).json("something wrong");
+  //     // return;
+  //     return next(error);
+  //   }
+  //   res.json(postArr);
 
   //validation
